Address review finding: move idempotent registration test to library suite
Review finding:

The new test lives in the shared abstract base class, so it also runs in the javaagent suite, where it cannot test what its name claims. In javaagent/src/test/.../CommonsPoolInstrumentationTest.java both configure(GenericObjectPool, String) and shutdown(GenericObjectPool) are empty overrides, because the agent registers metrics from the GenericObjectPool constructor advice and unregisters them from the close advice. In that suite the two configure(pool, ...) calls do nothing at all, ignoredPoolName is never handed to any registration, and verifyPoolNameNotReported(ignoredPoolName) can never fail. The javaagent copy of shouldRegisterSamePoolInstanceIdempotently therefore duplicates shouldReportGenericObjectPoolMetricsWhenJmxDisabled while advertising duplicate-registration coverage it does not have. Move the test, and the verifyPoolNameNotReported helper it needs, into library/src/test/java/io/opentelemetry/instrumentation/apachecommonspool/v2_0/CommonsPoolInstrumentationTest.java, where configure actually calls CommonsPoolTelemetry.registerMetrics twice, widening the base-class helpers it depends on (assertGenericObjectPoolMetrics) to protected as needed.

Analysis: Duplicate registration is only reachable through CommonsPoolTelemetry.registerMetrics, which the library suite calls from configure and the javaagent suite never calls, because the agent registers a pool from its constructor advice exactly once. The test therefore belongs in the library suite, which is where it can fail. It moves unchanged, keeping both configure calls and both assertions. assertGenericObjectPoolMetrics and verifyPoolNameNotReported become protected in the base class so the library suite can call them; verifyPoolNameNotReported stays in the base class so the POOL_NAME attribute key stays private there.

Upsides: The test now runs only where it exercises the behavior its name describes, so a passing run means duplicate registration really was ignored. The javaagent suite loses a test that could never fail and that repeated shouldReportGenericObjectPoolMetricsWhenJmxDisabled.

Downsides: Two base-class helpers widen from private to protected, which the javaagent suite can now call without needing to.
